% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BTlib.R
\name{bar_graph}
\alias{bar_graph}
\title{Draws a bar graph}
\usage{
bar_graph(data, xcol, ycol, outfile, bgfill = "white", font = "Arial",
  fill = NULL, colourtype = "discrete", x_axis_title = TRUE,
  text_size = 17, legend_pos = "right", extra = NULL, curve = NULL,
  segment = NULL, x_axis_hjust = 1, x_axis_vjust = 0,
  position = "dodge", barlabel = NULL, grid_color = "grey90",
  x_text_angle = 90, x_text_size = NA, manual_colors = c("#006077",
  "#0062D0"), facet_group = NULL, conditional_bold_x_labels = FALSE,
  x_axis_margin1 = 0, x_axis_margin2 = 0, x_axis_margin3 = 0,
  x_axis_margin4 = 0, continuous_x_breaks = NULL,
  continuous_y_breaks = NULL, ylim = c(NA, NA), xlim = c(NA, NA),
  height = NULL, width = NULL, x_grid_lines = FALSE)
}
\arguments{
\item{data}{A "melted" data frame (using the "melt" function from the reshape2 module)}

\item{xcol}{The name of the column to use as the bars}

\item{ycol}{The name of the column to use as the values}

\item{outfile}{The name of the output file}

\item{bgfill}{The background color of the plot}

\item{font}{The font to use in the plot}
}
\description{
Draws a bar graph
}
\keyword{bar}
\keyword{graphics,}
\keyword{plot}
