% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BTlib.R
\name{density_plot}
\alias{density_plot}
\title{Draws a line plot using ggplot}
\usage{
density_plot(data, column, fill_col, outfile, height = NULL, width = NULL,
  alpha = 0.1, ylab = NULL, xmin = NULL, xmax = NULL)
}
\description{
Draws a line plot using ggplot
}
\keyword{graphics,}
\keyword{line}
\keyword{plot}
