% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BTlib.R
\name{horizontal_line_graph2}
\alias{horizontal_line_graph2}
\title{Draws a horizontal line plot}
\usage{
horizontal_line_graph2(data, category_col, value_col, color_col, outfile,
  bgfill = "white", font = "Arial", x_axis_title = TRUE,
  y_axis_title = FALSE, y_axis_title_size = FALSE, text_size = 10,
  legend_pos = "right", grid_color = "grey90", image_height = 2,
  scale_color_manual = scale_colour_manual(values = c("#fb6a4a", "#73c475")),
  xlogscale = FALSE, x_text_angle = 90, x_axis_vjust = 0.5,
  x_text_size = 10, facet_group = "None", y_text_size = 10,
  vline_intercepts = NULL, vlinedf = NULL, num_x_ticks = 10,
  loglimits = c(NA, NA))
}
\arguments{
\item{data}{Data to plot}

\item{outfile}{The name of the output file}

\item{bgfill}{The background color of the plot}

\item{font}{The font to use in the plot}

\item{categories_col}{The name of the column to use as the categories}

\item{values_col}{The name of the column to use as the values}
}
\description{
Draws a horizontal line plot
}
\keyword{bar}
\keyword{graphics,}
\keyword{plot}
