% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BTlib.R
\name{int_with_commas}
\alias{int_with_commas}
\title{Formats an integer so that it has thousands commas}
\usage{
int_with_commas(x)
}
\arguments{
\item{x}{The integer to format}
}
\description{
Formats an integer so that it has thousands commas
}
\keyword{format,}
\keyword{integer}
\keyword{number,}
