% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BTlib.R
\name{line_or_scatter_plot}
\alias{line_or_scatter_plot}
\title{Draws a line plot using ggplot}
\usage{
line_or_scatter_plot(data, xcol, ycol, outfile, groupby = NULL,
  point_shapes = NULL, xbreaks = NULL, nxbreaks = 10, nybreaks = 10,
  bgfill = "white", font = "Arial", text_size = 15, ylab = NULL,
  pgmx = element_line(colour = "lightgrey", size = 0.15),
  pgmy = element_line(colour = "lightgrey", size = 0.15), box_x_min = 1,
  box_x_max = 0, rectfillcolor = "dodgerblue2", geom_line = FALSE,
  geom_point = TRUE, geom_point_alpha = 1, ylogbreaks = NULL,
  xrev = FALSE, ybreaks = NULL, manual_colors = NULL,
  vline_intercepts = NULL, hline_intercepts = NULL, ylim1 = NA,
  ylim2 = NA, xlim1 = NA, xlim2 = NA, geom_hexbin = FALSE,
  facet_group = "None", title = NULL, legend_title = FALSE,
  point_size = 1, geom_label = NULL, geom_label_size = 1.9,
  height = NULL, width = NULL, legend_position = "right",
  num_facet_cols = NULL, label_wrap_gen_width = NULL,
  facet_text_size = 15, jitter = FALSE)
}
\arguments{
\item{data}{A data frame}

\item{xcol}{The name of the column for the x axis}

\item{ycol}{The name of the column for the y axis}

\item{outfile}{The name of the output file}

\item{groupby}{Which variable to group by?}

\item{nxbreaks}{Number of x breaks}

\item{nybreaks}{Number of y breaks}

\item{bgfill}{The background color of the plot}

\item{font}{The font to use in the plot}

\item{pgmx}{X major gridlines}

\item{pgmy}{Y major gridlines}
}
\description{
Draws a line plot using ggplot
}
\keyword{graphics,}
\keyword{line}
\keyword{plot}
