% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BTlib.R
\name{scatterplot}
\alias{scatterplot}
\title{Makes a scatterplot}
\usage{
scatterplot(data, draw_best_fit = TRUE, draw_diagonal = FALSE,
  show_pearson_correlation = FALSE)
}
\arguments{
\item{data}{A two-column matrix or data frame, where the first column contains the x values and the second column contains the y values}

\item{draw_best_fit}{If true, draw a blue best fit line}

\item{draw_diagonal}{If true, draw a red line representing the diagonal}

\item{draw_diagonal}{If true, indicate the Pearson correlation between the two variables below the x axis label}
}
\description{
Draws a scatterplot.
}
\keyword{graphics,}
\keyword{scatterplot}
