% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BTlib.R
\name{scatterplot_ggplot}
\alias{scatterplot_ggplot}
\title{Draws a scatterplot using ggplot}
\usage{
scatterplot_ggplot(data, xcol, ycol, outfile, bgfill = "white",
  font = "Arial", pgmx = element_blank(), pgmy = element_blank())
}
\arguments{
\item{data}{A data frame}

\item{xcol}{The name of the column for the x axis}

\item{ycol}{The name of the column for the y axis}

\item{outfile}{The name of the output file}

\item{bgfill}{The background color of the plot}

\item{font}{The font to use in the plot}

\item{pgmx}{X major gridlines}

\item{pgmy}{Y major gridlines}
}
\description{
Draws a scatterplot using ggplot
}
\keyword{graphics,}
\keyword{scatterplot}
